v0.4.2 move templates.go to internal package
